removed dead code (refactor WIP)
